Make ExecutableLaunchConfiguration internal
