+++ src/Aspire.Hosting/Dcp/Model/Executable.cs
@@ -265,31 +265,6 @@ internal sealed class Executable : CustomResource<ExecutableSpec, ExecutableStat
     }
 }
 
-internal static class ExecutableLaunchMode
-{
-    public const string Debug = "Debug";
-    public const string NoDebug = "NoDebug";
-}
-
-/// <summary>
-/// Base properties for all executable launch configurations.
-/// </summary>
-/// <param name="type">Launch configuration type indicator.</param>
-public class ExecutableLaunchConfiguration(string type)
-{
-    /// <summary>
-    /// The launch configuration type indicator.
-    /// </summary>
-    [JsonPropertyName("type")]
-    public string Type { get; set; } = type;
-
-    /// <summary>
-    /// Specifies the launch mode. Currently supported modes are Debug (run the project under the debugger) and NoDebug (run the project without debugging).
-    /// </summary>
-    [JsonPropertyName("mode")]
-    public string Mode { get; set; } = System.Diagnostics.Debugger.IsAttached ? ExecutableLaunchMode.Debug : ExecutableLaunchMode.NoDebug;
-}
-
 internal class ProjectLaunchConfiguration() : ExecutableLaunchConfiguration("project")
 {
     [JsonPropertyName("launch_profile")]
